Updated authentication logic

+++ pages/api/documents/verification/email_authcode.ts
@@ -2,6 +2,7 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { sendVerificationEmail } from "@/lib/emails/send-email-verification";
 import prisma from "@/lib/prisma";
 import z from "zod";
+import { generateAuthenticationCodeURL } from "@/lib/api/emails";
 
 const bodySchema = z.object({
   email: z.string().email(),
@@ -19,14 +20,22 @@ export default async function handle(
     const code = authenticationCode as string;
 
     //Check verification code in database and delete it
-    try {
-      const response = await prisma.emailAuthenticationCode.delete({
-        where: {
-          code
-        }
-      })
+    const response =  await prisma.emailAuthenticationCode.findFirst({
+      where: {
+        code
+      }
+    })
+    if (response) {
+      //Delete the code from database if not invited
+      if (!response.invited) {
+        const response = await prisma.emailAuthenticationCode.delete({
+          where: {
+            code
+          }
+        })
+      };
       res.status(200).json({ message: "Verification successful" });
-    } catch {
+    } else {
       res.status(401).json({ message: "Unauthorized access" });
     }
 
@@ -45,36 +54,15 @@ export default async function handle(
       return;
     }
 
-    // Generate authcode
-    const authenticationCode = generateUniqueString(12);
-    await prisma.emailAuthenticationCode.create({
-      data: {
-        email,
-        code: authenticationCode,
-        linkId
-      }
-    })
-    const URL = `${process.env.NEXTAUTH_URL}/view/${linkId}?authenticationCode=${authenticationCode}`;
+    // Generate authcode URL
+    const URL = await generateAuthenticationCodeURL(email, linkId, false);
 
     await sendVerificationEmail(email, URL);
-    console.log(URL);
-    res.status(200).json({ authenticationCode });
+    res.status(200).json({ message: "Invitation send successfully" });
 
   } else {
     // We only allow GET and POST requests
     res.setHeader("Allow", ["GET", "POST"]);
     return res.status(405).end(`Method ${req.method} Not Allowed`);
   }
-}
-
-function generateUniqueString(length: number) {
-  const charset = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"; // You can customize this as needed.
-  let uniqueString = "";
-
-  for (let i = 0; i < length; i++) {
-    const randomIndex = Math.floor(Math.random() * charset.length);
-    uniqueString += charset.charAt(randomIndex);
-  }
-
-  return uniqueString;
 }
